+++ game/tiles.py
@@ -1,29 +1,28 @@
 import random
-from piezas import DATA
-
-TOTALTILES = 100
-
-
+from piezas import *
+class NoTilesInTheBagException(Exception):
+    pass
 class Tile:
     def __init__(self, letter, value):
         self.letter = letter
         self.value = value
-
-    def __repr__(self) -> str:
-        return str("[" + self.letter + "](" + str(self.value) + "p.)")
-    
-    def __repr__(self):
+        
+    def __repr__(self): 
         return f"{self.letter}:{self.value}"
-
-class Comodin:
+class Comodin(Tile):
     def __init__(self):
-        self.letter = '_'
-        self.value = 0
+        super().__init__(letter='?', value=0)
 
+    def select_letter(self, selection):
+        selection = selection.upper()
+        for tile in LETTER_VALUES:
+            if selection == tile['letter']:
+                self.letter = tile['letter']
+                self.value = tile['value']
 class BagTiles:
     def __init__(self):
         self.tiles = []
-        for i in DATA:
+        for i in LETTER_VALUES:
             for _ in range(i["quantity"]):
                 self.tiles.append(Tile(i["letter"], i["value"]))
         random.shuffle(self.tiles)
@@ -36,7 +35,5 @@ class BagTiles:
         return tiles_taken
 
     def put(self, tiles: list):
-        if len(tiles) + len(self.tiles) <= TOTALTILES:
-            self.tiles.extend(tiles)
-
-
+        self.tiles.extend(tiles)
+        random.shuffle(self.tiles)